:truck: move `InnerObservablesPool` in `Deployers`
