+++ src/lib/deployers/index.ts
@@ -4,6 +4,7 @@
  */
 export * from './instance-pool'
 export * from './instance-pool-worker'
+export * from './inner-observables-pool'
 export * from './models'
 export * from './in-worker'
 export * from './utils'
